requirements.txt
pandas
numpy
matplotlib
streamlit
pyyaml
sqlalchemy
mysql-connector-python